+++ src/client/render.tsx
@@ -2,7 +2,7 @@
 import { h, render } from 'preact';
 import { HttpRegion } from 'httpyac';
 import type { NotebookRendererApi } from 'vscode-notebook-renderer';
-import { TestResults, RFC7230Response } from './components';
+import { TestResults, RFC7230Response, HttpBody } from './components';
 
 interface IRenderInfo {
   container: HTMLElement;
@@ -17,9 +17,21 @@ export function renderCell({ container, mimeType, data }: IRenderInfo): void {
     if (data.testResults) {
       render(<TestResults testResults={data.testResults} />, container);
     }
-  } else if (mimeType === 'x-application/httpbook-rfc7230') {
-    if (data.response) {
-      render(<RFC7230Response response={data.response} />, container);
+  } else if (data.response) {
+    switch (mimeType) {
+      case 'x-application/httpbook-rfc7230':
+        render(<RFC7230Response response={data.response} requestVisible={true} bodyVisible={true} />, container);
+        break;
+      case 'x-application/httpbook-rfc7230-response':
+        render(<RFC7230Response response={data.response} bodyVisible={true} />, container);
+        break;
+      case 'x-application/httpbook-rfc7230-header':
+        render(<RFC7230Response response={data.response} requestVisible={true} />, container);
+        break;
+      default:
+        render(<HttpBody body={data.response.body} />, container);
+        break;
     }
+
   }
 }