+++ src/extension/httpOutputProvider/index.ts
@@ -1,0 +1,5 @@
+export * from './httpOutputProvider';
+export * from './contentTypeHttpOutputProvider';
+export * from './markdownHttpOutputProvider';
+export * from './rfc7230HttpOutputProvider';
+export * from './testResultsHttpOutputProvider';