+++ src/extension/mimeOutputProvider/index.ts
@@ -1,4 +1,0 @@
-export * from './contentTypeMimeOutpoutProvider';
-export * from './markdownMimeOutpoutProvider';
-export * from './markdownTestResultsMimeOutputProvider';
-export * from './rawMimeOutpoutProvider';