+++ src/extension/models/mimeOutputProvider.ts
@@ -1,6 +1,0 @@
-import { HttpRegion } from 'httpyac';
-import * as vscode from 'vscode';
-
-export interface MimeOutputProvider{
-  getNotebookCellOutputItem(mime: string, httpRegion: HttpRegion) : vscode.NotebookCellOutputItem | false;
-}