+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.0 (2021-05-06)
+
+#### Features
+* custom renderer added
+* support for RFC7230 output
+* extension api refactored
+* better settings support
+
 ## 0.1.0 (2021-05-03)
 
 #### Features